+++ BOSS/source/CombatSearch_BestResponse.cpp
@@ -28,7 +28,7 @@ void CombatSearch_BestResponse::recurse(const GameState & state, size_t depth)
     ActionSet legalActions;
     generateLegalActions(state, legalActions, _params);
     
-    for (UnitCountType a(0); a < legalActions.size(); ++a)
+    for (UnitCountType a(0); a < (int)legalActions.size(); ++a)
     {
         size_t ri = legalActions.size() - 1 - a;
